refactor(i18n): extract language detection logic
- move language resolution logic to a standalone utility function
- improve maintainability by separating detection concerns from i18n configuration
